moved the notification directory for cleanup, refactored file
